+++ CHANGELOG.md
@@ -12,6 +12,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 - Removed BC handler for deprecated `view` `_action`
 
+### Removed
+- The fallback mechanism that loads a default template when the template
+specified in a field description cannot be found was removed.
+
 ## [3.x]
 ### Added
 - Added AbstractAdmin, replacing Admin